+++ frontend/super-frontend/source/src/javascripts/super/flatpickr_controller.js
@@ -1,0 +1,14 @@
+import { Controller } from "stimulus";
+import flatpickr from "flatpickr";
+
+export default class extends Controller {
+  static get values() {
+    return {
+      options: Object,
+    };
+  }
+
+  connect() {
+    this.flatpickrInstance = flatpickr(this.element, this.optionsValue);
+  }
+}