+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add a time display type that only shows the time
 * Add a date-, datetime-, and time-picker to forms
+* Add the datetime-picker to filters
 
 ## [0.0.12] - 2021-04-04