+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add a date-, datetime-, and time-picker to forms
 * Add the datetime-picker to filters
 
+### Fixed
+
+* ActionText fields are computed
+
 ## [0.0.12] - 2021-04-04
 
 ### Added